// src/integrations/supabase/database.types.ts
export type Database = {
  public: {
    Tables: {
// ...
      clients: {
        Row: {
          address: string | null
          auditor_id: string
          created_at: string | null
          fiscal_year_end: string | null
          id: string
          name: string
          sector: string | null
          tin: string | null
        }
        Insert: {
          address?: string | null
          auditor_id: string
          created_at?: string | null
          fiscal_year_end?: string | null
          id?: string
          name: string
          sector?: string | null
          tin?: string | null
        }
        Update: {
          address?: string | null
          auditor_id?: string
          created_at?: string | null
          fiscal_year_end?: string | null
          id?: string
          name?: string
          sector?: string | null
          tin?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "clients_auditor_id_fkey"
            columns: ["auditor_id"]
            isOneToOne: false
            referencedRelation: "profiles"
            referencedColumns: ["id"]
          },
        ]
      }
      engagements: {
        Row: {
          audit_team: string[] | null
          client_id: string
          created_at: string | null
          end_date: string | null
          id: string
          start_date: string | null
          status: string | null
          year_end_period: string
        }
        Insert: {
          audit_team?: string[] | null
          client_id: string
          created_at?: string | null
          end_date?: string | null
          id?: string
          start_date?: string | null
          status?: string | null
          year_end_period: string
        }
        Update: {
          audit_team?: string[] | null
          client_id?: string
          created_at?: string | null
          end_date?: string | null
          id?: string
          start_date?: string | null
          status?: string | null
          year_end_period?: string
        }
        Relationships: [
          {
            foreignKeyName: "engagements_client_id_fkey"
            columns: ["client_id"]
            isOneToOne: false
            referencedRelation: "clients"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}